safari-600.7-branch/LayoutTests/fast/regions/cssom/webkit-named-flow-collection-empty-getters.html

<!doctype html>
<html>
    <body>
        <script src="../../../resources/js-test-pre.js"></script>
        <script>
            description("Test for <a href='https://bugs.webkit.org/show_bug.cgi?id=118561'>[118561] NamedFlowCollection getters should follow the same pattern as HTMLCollection</a>");

            var namedFlowCollection = document.webkitGetNamedFlows();
            shouldBeUndefined("namedFlowCollection[0]");
            shouldBeUndefined("namedFlowCollection['flow']");
            shouldBeNull("namedFlowCollection.item(0)");
            shouldBeNull("namedFlowCollection.namedItem('flow')");
        </script>
        <script src="../../../resources/js-test-post.js"></script> 
    </body>
</html>
